refactor: log 추가
